Global risk using log-linear models.
Source: R/LLmodGlobalRisk.R

The sample frequencies are assumed to be independent and following a Poisson distribution. The parameters of the corresponding parameters are estimated by a log-linear model including the main effects and possible interactions.

LLmodGlobalRisk(obj, method = "IPF", inclProb = NULL, form = NULL,
  modOutput = FALSE)

Arguments

obj
sdcMicroObj-class-object or a data.frame containing the categorical key variables.

method
At this time, only iterative proportional fitting (“IPF”) can be used.

inclProb
Inclusion probabilites (experimental)

form
A formula specifying the model.

modOutput
If TRUE, additional output is given.

Value

Two global risk measures or the modified risk in the sdcMicroObj-class object.

Details

This measure aims to (1) calculate the number of sample uniques that are population uniques with a probabilistic Poisson model and (2) to estimate the expected number of correct matches for sample uniques.

ad 1) this risk measure is defined over all sample uniques (SU) as $$ \tau_1 = \sum\limits_{SU} P(F_k=1 | f_k=1) \quad , $$ i.e. the expected number of sample uniques that are population uniques.

ad 2) this risk measure is defined over all sample uniques (SU) as $$ \tau_2 = \sum\limits_{SU} P(F_k=1 | f_k=1) \quad , CORRECT! $$

Since population frequencies \(F_k\) are unknown, they has to be estimated.

The iterative proportional fitting method is used to fit the parameters of the Poisson distributed frequency counts related to the model specified to fit the frequency counts. The obtained parameters are used to estimate a global risk, defined in Skinner and Holmes (1998).

Note

LLmodGlobalRisk is depcrecated for modRisk and is only provided for compatibility with older versions of this package. It may be removed in future versions.

References

Skinner, C.J. and Holmes, D.J. (1998) Estimating the re-identification risk per record in microdata. Journal of Official Statistics, 14:361-372, 1998.

Rinott, Y. and Shlomo, N. (1998). A Generalized Negative Binomial Smoothing Model for Sample Disclosure Risk Estimation. Privacy in Statistical Databases. Lecture Notes in Computer Science. Springer-Verlag, 82--93.

Templ, M. Statistical Disclosure Control for Microdata: Methods and Applications in R. Springer International Publishing, 287 pages, 2017. ISBN 978-3-319-50272-4. doi: 10.1007/978-3-319-50272-4

See also

loglm, measure_risk

modRisk