Adding noise to perturb data
Source: R/addNoise.r

Various methods for adding noise to perturb continuous scaled variables.

addNoise(obj, variables = NULL, noise = 150, method = "additive", ...)

Arguments

obj
either a data.frame or a sdcMicroObj-class that should be perturbed

variables
vector with names of variables that should be perturbed

noise
amount of noise (in percentages)

method
choose between ‘additive’, ‘correlated’, ‘correlated2’, ‘restr’, ‘ROMM’, ‘outdect’

...
see possible arguments below

Value

If ‘obj’ was of class sdcMicroObj-class the corresponding slots are filled, like manipNumVars, risk and utility.

If ‘obj’ was of class “data.frame” or “matrix” an object of class “micro” with following entities is returned:

x

the original data

xm

the modified (perturbed) data

method

method used for perturbation

noise

amount of noise

Details

If ‘obj’ is of class sdcMicroObj-class, all continuous key variables are selected per default. If ‘obj’ is of class “data.frame” or “matrix”, the continuous variables have to be specified.

Method ‘additive’ adds noise completely at random to each variable depending on its size and standard deviation. ‘correlated’ and method ‘correlated2’ adds noise and preserves the covariances as descriped in R. Brand (2001) or in the reference given below. Method ‘restr’ takes the sample size into account when adding noise. Method ‘ROMM’ is an implementation of the algorithm ROMM (Random Orthogonalized Matrix Masking) (Fienberg, 2004). Method ‘outdect’ adds noise only to outliers. The outliers are idedentified with univariate and robust multivariate procedures based on a robust mahalanobis distancs calculated by the MCD estimator.

References

Domingo-Ferrer, J. and Sebe, F. and Castella, J., “On the security of noise addition for privacy in statistical databases”, Lecture Notes in Computer Science, vol. 3050, pp. 149-161, 2004. ISSN 0302-9743. Vol. Privacy in Statistical Databases, eds. J. Domingo-Ferrer and V. Torra, Berlin: Springer-Verlag. http://crises-deim.urv.cat/webCrises/publications/isijcr/lncs3050OntheSec.pdf,

Ting, D. Fienberg, S.E. and Trottini, M. “ROMM Methodology for Microdata Release” Joint UNECE/Eurostat work session on statistical data confidentiality, Geneva, Switzerland, 2005, http://www.unece.org/fileadmin/DAM/stats/documents/ece/ces/ge.46/2005/wp.11.e.pdf

Ting, D., Fienberg, S.E., Trottini, M. “Random orthogonal matrix masking methodology for microdata release”, International Journal of Information and Computer Security, vol. 2, pp. 86-105, 2008.

Templ, M. and Meindl, B., Robustification of Microdata Masking Methods and the Comparison with Existing Methods, Lecture Notes in Computer Science, Privacy in Statistical Databases, vol. 5262, pp. 177-189, 2008.

Templ, M. New Developments in Statistical Disclosure Control and Imputation: Robust Statistics Applied to Official Statistics, Suedwestdeutscher Verlag fuer Hochschulschriften, 2009, ISBN: 3838108280, 264 pages.

Templ, M. and Meindl, B. and Kowarik, A.: Statistical Disclosure Control for Micro-Data Using the R Package sdcMicro, Journal of Statistical Software, 67 (4), 1--36, 2015. doi: 10.18637/jss.v067.i04

Templ, M. Statistical Disclosure Control for Microdata: Methods and Applications in R. Springer International Publishing, 287 pages, 2017. ISBN 978-3-319-50272-4. doi: 10.1007/978-3-319-50272-4

See also

sdcMicroObj-class, summary.micro

Examples

data(Tarragona)
a1 <- addNoise(Tarragona)
a1

#> $x
#>     FIXED.ASSETS CURRENT.ASSETS TREASURY UNCOMMITTED.FUNDS PAID.UP.CAPITAL
#> 1              0              0     6670            112349           37500
#> 2              0           6643     1892            -12126             500
#> 3              0          43279     6863             29125           34000
#> 4              0          98448        1              2712             500
#> 5              0         134917    11368             19893           10000
#> 6              0         162855      507            117733           30000
#> 7              0         170531     2274             -6118             500
#> 8             62           1389       97              1049             500
#> 9             78          26613     3490             13242           10000
#> 10            91         183865    22670             34555           10000
#> 11           140          10225     5281              2050             500
#> 12           142          75370    28605             55605            4000
#> 13           200          54989    25592             16084            6000
#> 14           224          31228     6951             20331            4000
#> 15           237            714        1               212              90
#> 16           249          65547    20558             50769           48000
#> 17           310         195474    38990             31160           10000
#> 18           323          63115     4916             54986           15000
#> 19           391          24611     1355              4324            1000
#> 20           408          29619      775              5395             500
#> 21           435           7807     7878               831           10000
#> 22           436          31314     2317             23651           16000
#> 23           465         354505    17608             42964           10000
#> 24           474          40515      736              3074             500
#> 25           684          16787        0              6303            6000
#> 26           893          10819    10106              6474           10000
#> 27           905          52413     1038             24818            6000
#> 28           989          45102     1369             17772           30000
#> 29          1000         615990      124           -515464           10000
#> 30          1013           1263   -12228            -10106             500
#> 31          1048          33783    12934             22042           10000
#> 32          1062          18946    19649              5268            2000
#> 33          1120          83212     2727             53973           15000
#> 34          1128          49746      584              9495             500
#> 35          1306          45305    12553             36218             725
#> 36          1307          12761      469              1381           10000
#> 37          1327          75294    30101             49158           10000
#> 38          1381          38433    20625             39598           36500
#> 39          1426          43934     3647             11229           10000
#> 40          1428          42005      963             26821           19260
#> 41          1465         369131     1721             79189          140000
#> 42          1489          68099       30            -26447            2000
#> 43          1490         188830    25095             16466            2000
#> 44          1496          39567       81              4023           15000
#> 45          1747          22079     9403              9526           10000
#> 46          1784         322333    11200             38217           10392
#> 47          1816         102019     6404             33755           12000
#> 48          1835          88524      374              4639             500
#> 49          1968          81069     1606             19452           11000
#> 50          2045          21484     7253             18694            2000
#> 51          2090          33917     2324             10031            1000
#> 52          2126          35771     1509             11687            7000
#> 53          2136           7906     4040              9686            7000
#> 54          2402          49630      310             33617           32000
#> 55          2429         207875       83             31867           23000
#> 56          2538          17093     2662              6273            1000
#> 57          2567         117238      671             14095            7850
#> 58          2739          23350        0             14778           10000
#> 59          2809          89592    13926             14141           10000
#> 60          2830          30250     5922              6424            3000
#> 61          2860          68515     1442             13120            2000
#> 62          2871          48713     9856             39106            1500
#> 63          2885         158347    16306             48499           10000
#> 64          3062          70586      465              6832           10000
#> 65          3243          98781    34576             31758           20000
#> 66          3274          67444     2216             15834            2000
#> 67          3452          53685     1252             20658            3000
#> 68          3487         134897      695             62728           10000
#> 69          3506          42499    23167             29732            6500
#> 70          3514         101242     7152             28771           15000
#> 71          3543          28253     2389              9413            1500
#> 72          3544          35812       81              8967             500
#> 73          3647              0        0             -1142             500
#> 74          3787          24615      985             10859            5000
#> 75          3819         107416     1400             27300           15000
#> 76          3822          39864     3993              3595            1000
#>     SHORT.TERM.DEBT    SALES LABOR.COSTS DEPRECIATION OPERATING.PROFIT
#> 1            183077   507061       60067         9006            27560
#> 2             13872    44727        5090            0             -770
#> 3             14154    92922        8583            0            -6434
#> 4             94684   665156        2663            0             3658
#> 5             40669    86020       27054            0            27468
#> 6             35180   297633       35086        19110            15194
#> 7            176649   378474        5105            0             5688
#> 8               402     7849        2376            0              243
#> 9             20948     3430        3112            8              103
#> 10           149402   637815        4085            0             8504
#> 11             8314   371017        5597            0              546
#> 12            19907   233917      179835          381             9704
#> 13            39105   139150      122103          210             1597
#> 14            11120    74373        4784          577             5115
#> 15              739     1857         201            7               64
#> 16            15027   339469        3548          110            -1565
#> 17           164623   441054       23564            0             6946
#> 18             8452   125582       18524          183             3884
#> 19            20678   380017         591           53             4750
#> 20            21333   233756       14679          372             2873
#> 21             7411   224768       23275          214             3947
#> 22             7579    69645        4357          141              878
#> 23           229572   116295           0          362             -593
#> 24            37915    46159         377          692            11871
#> 25            11167   116534        4253          805             3603
#> 26             4853    77844       10487           37               39
#> 27            28500   143403       11578          308            -4965
#> 28            28320    84497       11363          453            12563
#> 29           687142  2045712     1338127            0           -45996
#> 30             4638    34752        3290           92            -2101
#> 31            15789   378108       29709          559             5999
#> 32            14740    31969       14408          301              112
#> 33            10859     8317        3615          307             1539
#> 34            34103   100068       62352          271             1065
#> 35            10393    96500       21473         3118            10932
#> 36            12687   210422        5293            0            -8155
#> 37            27463   133921       13424          248             4436
#> 38              216   600266        4758          388             -456
#> 39            11931   100932        8865          286             2566
#> 40            16612    96499       12566          311             1620
#> 41            18128    94412         351          400            14395
#> 42            71036     3773       17177            0           -14777
#> 43           173854   726488       37010          497             8120
#> 44            22686   166380       20050          631             2762
#> 45            14300   130326       10037          183            -1300
#> 46            80751   130349       12034          128            11187
#> 47            70079   141827       23987          684             2004
#> 48            85720   135138       10738            0             4926
#> 49            63585   136766       17006          569             7843
#> 50             4359   127285       15007          251             3287
#> 51            25975   112162       16746          780             1057
#> 52            26199   118270       12044          465              606
#> 53              356    49335        4414          608              452
#> 54            18495    70200        3959          787             -360
#> 55           162415   292284       28000          817            14289
#> 56            12092   157843        6297          622            -2100
#> 57           105710   310463        9161          905             1607
#> 58            11310    99855        7822          618             2509
#> 59            46803    90834        9553          157             5701
#> 60            26657   133264       22891         1789             4305
#> 61            58254   207609        5835          579             2917
#> 62            10393    96519       19928         1561             3239
#> 63            90529   289985       59491         1526           -10360
#> 64            66816   141618       45310          184             -607
#> 65            70266   204889       10628          859            11448
#> 66            44884   113000       19826         1135             8773
#> 67            35581    86259       31731         1887            13633
#> 68            75656   352846       35952         1311            27396
#> 69            16273    52310       10444          107             1801
#> 70            18302   127663       26272          626            -2003
#> 71             3304   159372        7847          985             4052
#> 72            30389   135055        5176          697             1411
#> 73             4789    46891       16756            6            -3644
#> 74            17070    81321       14588          740              -15
#> 75            70662   168561        8741          572             8567
#> 76            38396   143473       16937         1404             1366
#>     FINANCIAL.OUTCOME GROSS.PROFIT NET.PROFIT
#> 1              -18931         9218       6338
#> 2                 -60         -830       -830
#> 3                 706        -5727      -5727
#> 4               -6625         -944       -685
#> 5              -25480         2580       1496
#> 6                 725        16450      10692
#> 7              -13442        -7753      -7753
#> 8                  -9          234        152
#> 9                 281          664        432
#> 10                  0         3117       1955
#> 11                500         1047        857
#> 12                999        10703       6957
#> 13               1218         1203        782
#> 14                 89         5205       3307
#> 15                -43           -5        -14
#> 16               2165          600        390
#> 17              -5028         1947       1272
#> 18              -2741         1142        742
#> 19              -3563         1178        771
#> 20              -2817          263        107
#> 21              -3353          594        386
#> 22               -387          491        319
#> 23             -11552       -12615      -8200
#> 24              -1324        10682       7141
#> 25              -1013         2589       1683
#> 26                 15         2822       1834
#> 27               -212        -5074      -5074
#> 28                 30        12866      12866
#> 29             -18660        26007      26007
#> 30               -983        -3045      -3005
#> 31              -4882         1116        726
#> 32               1336         1448        963
#> 33                 75         1614       1049
#> 34              -3179        -2114      -2114
#> 35                838        11771       7651
#> 36               -145        -8281      -6481
#> 37               2819         7855       5134
#> 38               1072          615        448
#> 39              -3573        -1006       -654
#> 40                -58         1561       1561
#> 41             -32081       -17643     -17643
#> 42                -28       -14806     -14806
#> 43              -2574         5546       3642
#> 44              -2555          207        134
#> 45               -242        -1542      -1520
#> 46             -24580       -13393      -8750
#> 47                -15         5584       3647
#> 48              -4514          411        411
#> 49              -6960          882        882
#> 50                 83         3253       1991
#> 51               -866          147         95
#> 52               -799         -193       -125
#> 53                  0          452        312
#> 54               1301          992        992
#> 55             -12682         1953       1269
#> 56               -888        -2988      -1028
#> 57              -4338        -2700      -1755
#> 58                488         2985       1940
#> 59              -4298         1402        911
#> 60               1941         2364       2364
#> 61              -2956        -2956      -1921
#> 62                844         4052       2633
#> 63               -950         2011       1307
#> 64              -7615          141        141
#> 65               1984        13600      10127
#> 66              -6830         1852       1180
#> 67              -3478         2823       2823
#> 68              -5091        22389      14552
#> 69               1711         3512       2342
#> 70              -5661         1723       1163
#> 71              -3754          298        193
#> 72               -802         1218        759
#> 73              -2039        -5684      -5684
#> 74               -137          250        182
#> 75              -6658         1909       1240
#> 76              -1103          262        184
#>  [ erreichte getOption("max.print") --  758 Zeilen ausgelassen ]
#> 
#> $xm
#>     FIXED.ASSETS CURRENT.ASSETS      TREASURY UNCOMMITTED.FUNDS PAID.UP.CAPITAL
#> 1    -584296.542    -134825.517  142179.65198      -128567.2399     -72709.3439
#> 2     106554.454     281626.002 -121404.01650     -1054458.1644      91811.1444
#> 3   -1017171.739    -285360.670   18459.73097       872407.7822       -524.5995
#> 4      -2325.130    -298180.844  -51180.29132      -281242.9922     483422.7627
#> 5     259399.869    1699098.904  -63635.31963        99833.2935    -243047.4898
#> 6     479280.052      55864.548   80840.35331      -503066.2209     -77104.9133
#> 7    -760320.480    -115112.634   68070.39413      -617703.6184     116330.4496
#> 8    -103157.162     339938.326   80012.70212       431614.4867    -182578.4385
#> 9    -101836.679   -1217594.571   45926.22628       -39365.7638    -172953.3156
#> 10   -117893.773   -1102064.980   58399.96693      -160624.4924     -10997.6802
#> 11   -230941.842     348271.009  -49714.31039        38245.5427     -91876.3012
#> 12    262642.435     579833.599   17921.02120       513078.3952     268125.8634
#> 13    862021.001    -471744.423   24416.34002       813574.6730     191683.6466
#> 14   -680455.890     621931.650  -19245.95827       -33922.9273     126398.4461
#> 15    214094.134     616067.782  -27008.08514       747467.3782     -25842.5866
#> 16   -777263.383    -808786.461  106170.75554      -545125.7382     250745.0893
#> 17   -217547.590     425189.463   94049.27342       196020.5940      38758.0546
#> 18    -21629.971     407039.697   33065.73528       677547.2896     249397.7720
#> 19    227006.017    -177084.709  116134.69327      -154087.1820    -240245.8236
#> 20   -381073.542    -128738.211 -147799.75320      -242540.6564      48894.3812
#> 21    195815.362    -181502.725   -4519.46653        -8495.9177      27085.4520
#> 22    151910.694    -700353.135   47323.85309      -130368.2298    -178720.7932
#> 23   -543975.421    -373786.490  -35712.10330       127322.1574      81306.9036
#> 24    308378.824    1372682.213 -104293.30822       160280.7803      81274.4762
#> 25    788835.858    -573118.951   81389.84571      -354294.5543      61677.7406
#> 26    -39774.906     308119.840  110947.71856      -333805.4805     130542.8341
#> 27   -389663.126      22176.667  -77380.69852      -814557.1431      52252.6249
#> 28     -5667.730    -396559.391   42500.52901      -689859.6242      33634.2789
#> 29   -344053.508    -349652.146   99632.69940      -514393.5963     -23091.1473
#> 30   -630174.442    -253317.614  -41986.34976       658154.6275     -41409.6799
#> 31    391414.064   -1101045.943   30944.95202      -331352.1339    -305443.2059
#> 32     74718.057      -6133.499  -62712.12920       325539.6653     -54855.4174
#> 33    102820.106    -433814.845   34815.93437       370605.1853     -57225.4059
#> 34    678702.651    -544312.306  112560.58002       433109.5777      45994.1152
#> 35     48064.164      92825.657   10108.38290      1225363.5168     -45249.5115
#> 36    -54615.541     815828.434  -29099.07651      -670191.3675      91383.9558
#> 37   -795832.151     541223.567   79348.61170        83101.4351    -165570.0547
#> 38   -115156.345     388326.401  -17953.80424      -251654.6519       1130.3640
#> 39   -129388.077    -644373.385    8669.44145        63466.6786    -220866.9202
#> 40    446860.085    -308713.002    4939.63208       366368.5369      82741.5844
#> 41     30693.463     -94744.723   58603.49595      -429199.2838     367118.0622
#> 42   -265243.815     -47947.053  172729.61296       128551.7213     -74652.0625
#> 43    -19362.436     295828.325   31415.83591        73297.0856     159772.5228
#> 44   -103458.528    -432223.778  -34785.50692       145551.5238    -170008.2363
#> 45    187379.385    -326757.608   61334.23415       203359.7544     267677.3819
#> 46   1151734.655     -66604.580   93565.98618      -538851.5929      -6690.3941
#> 47     21235.485     868711.110  132512.76430       429114.8115    -110926.4769
#> 48    242937.085     211245.869   -8024.39473        -6076.5012    -116555.5701
#> 49     51295.650     242535.288  -49296.98440       313112.8186     140683.2257
#> 50   -795795.680     700328.055  -26899.61774       578987.3690     352244.6154
#> 51    361874.638    -802312.672   -8139.62745      -571960.7668    -105684.7844
#> 52    -99386.834     279067.783   10399.64488       -68350.3767     -60025.4408
#> 53    -83872.780    -353306.553  -61279.62277       531907.1857    -116648.6322
#> 54     10405.609     245781.352  -83575.53581        81983.8378     121275.5059
#> 55     14765.935     352035.148   -3024.17835       510310.5458    -330340.1053
#> 56    232003.961    -606960.346   78932.36120        35139.9304    -109558.7220
#> 57   -946515.961     150985.754   85461.82710       359673.0237     -53127.7961
#> 58   1122281.976     607108.572  -36797.26243      -250393.8988     337276.4748
#> 59   -147943.693    -166537.368   59685.45177      -364567.3291    -166150.9322
#> 60     91872.251     649101.015  -56426.26155      -427784.1266    -148977.6983
#> 61    451229.337    -312995.846  -63949.01696       225932.2624     -99076.1602
#> 62   -274698.060    -908987.223    8037.59697       716073.3937     123401.1468
#> 63    467783.796    -735266.700   46258.69243       432871.2168     119214.4330
#> 64    -99560.827    -802205.670  -21604.43903       -34226.3351     283677.0114
#> 65   -488203.273    -378996.146   42156.78673      -314131.5735     155343.4177
#> 66   -403989.155    -237466.336   72708.79731      -176299.6477      70347.9661
#> 67    447944.833     949136.528   23178.28147      -487053.6146    -164770.8720
#> 68     58438.647    1221294.907   53471.75003      -561442.3199       2705.7208
#> 69    207431.178     805830.679   52776.87664      -181457.4561     -14466.8545
#> 70   -705737.589       4667.800  -61199.99517      -313031.5181     -35789.1693
#> 71   -610256.591     -87081.729   -8252.91032      -176025.8239      76795.1722
#> 72    122131.788     556710.038 -109739.12666        89502.7291     129675.9786
#> 73    561766.575    -269681.343  190394.04718       453087.0814      41296.9130
#> 74    102570.229     811025.480   65482.17200        63707.4792    -336616.5291
#> 75    553997.802     911903.016   35494.17029      -806801.9297     124715.2998
#> 76    222471.401     484587.571   21395.97208      -132172.7261    -280787.8522
#>     SHORT.TERM.DEBT        SALES  LABOR.COSTS  DEPRECIATION OPERATING.PROFIT
#> 1       530382.5150  1261234.786 -139793.9037  8.404207e+04       65118.3543
#> 2       735093.6833 -1256366.023   68015.8097 -6.657441e+03      -59170.8497
#> 3       711898.5968   405482.830 -181067.8357  6.232456e+03      251820.8757
#> 4      -264074.4403  1715307.178  220745.9122  1.748045e+04     -113157.6985
#> 5      -561314.7432 -1249097.504  132063.7201 -8.844370e+04      -22333.8398
#> 6      -379861.3939  4887039.868 -104966.9187 -4.594361e+04       21574.7631
#> 7       981416.1164  1612371.159 -252435.1074  2.172603e+03       31141.5553
#> 8     -1243403.8697  -270799.289   90058.3390 -3.095151e+04     -102486.2297
#> 9       763690.3930  2808993.026  215151.6393  7.887850e+03      110646.4274
#> 10      808552.6511  1429307.564  264364.8097  4.216462e+04      -96549.8706
#> 11      176093.1286  3334495.094  -59289.9832  8.980214e+03      -96526.6039
#> 12     -139269.1134 -3909971.823  262703.6914  3.091657e+04      113754.5445
#> 13     -427102.9220 -1422064.173  -27050.6229  6.720689e+04        3224.2710
#> 14       71225.2679  -871787.604  -52376.2234  9.159695e+03      -44123.8605
#> 15     1084037.8211   793881.134  -29657.6877 -1.591533e+04     -325215.6978
#> 16      654793.7331  1575888.731   42079.1989  5.954387e+04      172205.4047
#> 17      -93020.6462  -547626.183  156091.1171  2.758640e+03     -163533.4037
#> 18      573064.2479  1258736.762   -1933.5025  1.501139e+04     -152470.5458
#> 19      425722.6321  -507283.937  175168.6442  2.091830e+04      147516.9218
#> 20      588204.8567     3970.805 -315683.8868  2.923102e+04      -87684.2706
#> 21      208898.4773   622585.688  132704.5330  3.192100e+04      207803.1830
#> 22      -19899.9541  -670644.577  370332.2819  3.529267e+04        2664.4023
#> 23      612069.2435 -1875285.713  -47156.8249 -1.845224e+04      -26469.9649
#> 24     -904960.8781 -1039247.388  154321.0669  5.385899e+04      -23070.5481
#> 25     -122995.0642  1504328.613  116427.1767  5.331807e+04      168141.2275
#> 26      300837.5372 -1610188.901 -164991.1053  4.988906e+04      -53278.8335
#> 27     -617356.0192 -1087793.690  -59029.8271 -1.952619e+04      324139.3253
#> 28      496816.1922   115860.605 -228251.9362 -2.307741e+04     -130921.7113
#> 29      693443.7008  -459580.282  854088.9839 -3.461175e+04     -122832.0744
#> 30      147520.2989  1662217.437   -5882.2452  4.588763e+04      -45745.4114
#> 31     -972989.7431   431486.295  -12059.2850 -1.659564e+04       79287.3509
#> 32     -767969.9595  1537473.301 -111313.6088 -1.190551e+04      -49343.2031
#> 33      398423.7537 -1009109.590 -435065.4119 -2.741269e+04      127113.4223
#> 34        6670.5590  -285404.200  298199.7580 -6.221838e+04     -210312.8173
#> 35     1375625.8455  -144839.216 -173735.8552 -5.130552e+03       61644.9035
#> 36      206545.3947  1383540.303 -116142.2227  6.988755e+04      -66757.7436
#> 37     1350192.3420  2035028.451 -396930.9424  1.287578e+04     -154822.1589
#> 38      518239.9785   590901.031  -96794.5845 -3.733780e+04       86400.5930
#> 39      885527.6098  2235005.021  482520.1815 -9.087288e+02       51997.8811
#> 40      663124.8460  -113192.714  -41604.1266 -4.375384e+03       76784.8421
#> 41     -336868.0618  -842696.112 -140296.6969  5.240082e+03      -62568.1610
#> 42      -58047.6922  1673629.468  149403.5964 -5.113042e+04      209102.9584
#> 43        2950.5001  -565646.242  280413.5723  1.102500e+04     -136116.0754
#> 44      -81864.9545 -4058684.018   96973.9220  9.525597e+03       66735.1753
#> 45      381348.5161  2330815.312   -3647.7605 -2.238191e+04     -157705.9346
#> 46    -1392288.8755  -406389.881  125857.7241  9.504247e+03      -93139.6702
#> 47      -30774.4845  -489378.738  429394.2072 -1.758481e+04       -6190.2645
#> 48       15358.0954 -1559875.619   -4716.5630 -5.958136e+04      169383.5070
#> 49      706307.6481  1276329.632  181131.2035  9.233589e+03     -126669.5662
#> 50      379206.1361 -1245613.318  -89863.3681  4.085690e+04      -24335.3513
#> 51      302716.8619  3224345.430   27197.3366  1.769464e+04       92440.3380
#> 52      -59755.7704  3812517.748   54427.8432 -2.275079e+04       73669.7103
#> 53        -906.2904 -1040908.177 -142258.9116  3.836700e+04     -140802.2564
#> 54      307141.2809  1019976.935  372178.1459  5.547591e+04       15227.4652
#> 55      248367.6980 -1583000.890 -503095.0681 -3.979787e+04      130259.2131
#> 56      696961.0662   261663.832  312065.3466 -2.705562e+04     -151010.1568
#> 57      510386.8079  1184471.071  -46370.3807 -3.881334e+04       22843.7908
#> 58      283277.5850 -1265102.979  -93996.9934  2.004562e+04     -172441.6973
#> 59      314185.3235 -2663309.574  -67109.6825  4.249373e+04     -159237.3096
#> 60     -236316.0929  2079220.032   94636.8675 -8.153163e+04      -61319.0909
#> 61     -843483.2091  1663968.908  120218.1282  1.738900e+04     -121219.8107
#> 62     -338846.8244   225931.276  362487.0174  8.216985e+03        -365.1483
#> 63       44165.9281 -4565790.551  -55596.8453 -1.210929e+04     -166373.7603
#> 64      185959.3766 -2808926.593  -36952.5495  1.222503e+05       37050.8280
#> 65      570559.8502  1775982.902  -95206.3967 -5.598375e+04      -50618.4210
#> 66      497531.4194   781958.173  306324.0281 -5.874004e+04      120357.2194
#> 67      164610.1618 -2056661.088  147686.3701  1.648043e+02      124778.5276
#> 68     -261446.7039 -2757056.006  250132.9037 -3.432977e+04       98075.3316
#> 69      124554.5540   326931.028  145916.4630  1.276047e+04       72142.8159
#> 70     -253358.3166  -823946.507  107710.7262 -1.555661e+04      177441.7155
#> 71      273857.5869    20292.334 -179614.8100 -5.238837e+04       28101.2950
#> 72      -48952.8161  -943838.663  149212.1020  4.281414e+04      184484.4076
#> 73      667341.2638 -1203149.748   11606.3201 -1.585703e+04       47369.1451
#> 74       14455.3217   830540.511  176849.5473  4.809929e+03       29486.1504
#> 75    -1108502.1541  1894617.713 -369116.0109 -1.603940e+04       86974.2502
#> 76      523669.5801  -768415.395  158239.9504 -1.474725e+04      -70481.7140
#>     FINANCIAL.OUTCOME GROSS.PROFIT   NET.PROFIT
#> 1        -41202.11087  115726.4148   53617.6102
#> 2          5848.08397   44833.5735   73737.9523
#> 3        -52647.53632 -366782.1846  -32474.9930
#> 4         62792.18598  -46501.4997   31740.6698
#> 5        -18351.66321 -139530.0764  -80455.1846
#> 6          7058.77353 -141856.7785 -104695.6638
#> 7         28140.86843  -10308.6929   67791.6512
#> 8         -1677.95870   38940.2203  -52492.1211
#> 9          6988.97668  135069.3202  -65248.5106
#> 10       -12481.34788  153811.3770  109742.5725
#> 11       -27532.07797 -219456.5882 -119573.3596
#> 12        41251.33026   37873.1587   40401.6349
#> 13        50377.20686   86815.4141  -59668.4465
#> 14       -40253.38709  -96738.8035 -224973.5009
#> 15        46048.09046    -219.0810  235489.2018
#> 16        33640.78539   11097.6439   32348.6104
#> 17       -23485.23754   44931.1255  -35632.9198
#> 18       -25417.64848   17798.2527   -4545.2827
#> 19        11508.62620   88053.5497   -2169.5190
#> 20        15416.10526  -15526.0472  -18161.3612
#> 21       -20813.96287 -123507.3661  -18915.4169
#> 22        -5299.41702  174266.2982 -139584.9365
#> 23        40353.10624  144625.9011  -73576.2496
#> 24        21462.66059  112259.1714  -28293.0242
#> 25        11563.89204 -116903.0260   18532.1200
#> 26        13560.95178  128242.1252  -11382.7383
#> 27      -105095.53574   35013.2061  -32615.3221
#> 28       -58264.40667    3006.1523   50647.0575
#> 29       -17110.26859   95061.5676   33799.7538
#> 30       -67181.21310 -212926.9593 -205983.2612
#> 31       -13156.89239 -113796.0046   -6541.1234
#> 32        -6162.25859  127338.9472  -17355.5905
#> 33        58194.98794  -94985.4044   29779.8103
#> 34        29732.85137  -70666.4824 -151793.6857
#> 35        28134.11680  130961.5918  -17007.0613
#> 36        43554.93057    3759.1908   94045.5771
#> 37        74379.00907  -31849.7996  -44193.0174
#> 38        47057.40119  -97011.3699   69489.8736
#> 39        21105.41074  -93609.5408  -63085.6923
#> 40        -7655.88650 -125284.7689  -30731.1229
#> 41         2323.42989 -263565.5730   20790.0184
#> 42       -50796.99355  -83369.1625   34306.2601
#> 43        32608.75221  -33584.0895   52075.0482
#> 44        14704.76272  -22387.7106  -77514.3670
#> 45         3018.52608  -42913.8338  -16006.6424
#> 46        24705.28589  -76626.2841 -136651.0555
#> 47        -2295.42240   99774.2368   18970.3539
#> 48        36327.23002   83243.6176  110283.3946
#> 49        32264.06266   48997.7340  -52707.5012
#> 50       -43541.99554 -104569.2312  -12068.8955
#> 51       -32828.70942  -93606.7015   77288.7860
#> 52       -29617.65192   16555.5646  -33230.0190
#> 53       -19676.93139 -130973.3381  169557.7378
#> 54       -64261.07640   28337.2079  -46819.1033
#> 55       -22844.85990  280674.7514 -114368.6319
#> 56       -37609.39267  229508.4502  153030.1249
#> 57       -69532.03386  -85453.4429   46709.7868
#> 58       -39120.93180 -120957.6131  -67697.3013
#> 59       -12579.04514  -69796.5590  -69363.2788
#> 60       -60262.59491   47927.2869  -64311.2000
#> 61       -26984.12716 -117273.2482  -32650.6783
#> 62        58088.71454  -42631.9877  -10999.7270
#> 63        37580.10655  188221.7057    8863.5180
#> 64        37573.35099  147502.3724   13760.0845
#> 65        56850.03957  169939.3974    2521.5037
#> 66        37514.91561 -109707.5014 -108791.7003
#> 67        17416.74864  -40213.0363  -29918.4690
#> 68        15531.57017  121667.7451   -9734.3477
#> 69        37342.61933    9920.0337    6179.4261
#> 70        31709.72351  -19743.0513    9711.8887
#> 71       -52685.84636 -120845.7055  108342.7751
#> 72        31059.03291  106846.9071 -130965.1008
#> 73       -33456.77860  -64442.3054   26505.9544
#> 74        42566.51419  120385.1239  -88912.0323
#> 75      -113163.50618 -107141.4500   90025.6204
#> 76         4397.98600   32335.9558  207677.0316
#>  [ erreichte getOption("max.print") --  758 Zeilen ausgelassen ]
#> 
#> $method
#> [1] "addNoise: additive"
#> 
#> $noise
#> [1] 150
#> 
#> $p
#> [1] NA
#> 
#> $delta
#> [1] NA
#> 
#> attr(,"class")
#> [1] "addNoise"

data(testdata)
testdata[, c('expend','income','savings')] <-
addNoise(testdata[,c('expend','income','savings')])$xm

## for objects of class sdcMicroObj:
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- addNoise(sdc)