Demo data set from mu-Argus

The public use toy demo data set from the mu-Argus software for SDC.

Format

The format is: num [1:4000, 1:34] 36 36 36 36 36 36 36 36 36 36 ... - attr(*, "dimnames")=List of 2 ..$ : NULL ..$ : chr [1:34] "REGION" "SEX" "AGE" "MARSTAT" ...

Source

Public use file from the CASC project.

Details

Please, see at the link given below. Please note, that the correlation structure of the data is not very realistic, especially concerning the continuous scaled variables which drawn independently from are a multivariate uniform distribution.

Examples

data(free1)
head(free1)
#>      REGION SEX AGE MARSTAT KINDPERS NUMYOUNG NUMOLD AGEYOUNG EDUC1 EDUC2 ETNI
#> [1,]     36   1  43       4        3        0      0       97     4     0    1
#> [2,]     36   1  27       4        3        0      0       97     4     0    9
#> [3,]     36   1  46       4        1        0      0       97     7     7    1
#> [4,]     36   1  27       4        1        0      0       97     5     0    1
#> [5,]     36   2  24       4        1        0      0       97     4     0    1
#> [6,]     36   2  24       4        1        0      0       97     4     0    1
#>      PRIOCCU POSLABM REGJOBC RECBEN RECUNBEN RECODBEN RECBILL RECSOSEC RECPENS
#> [1,]       1       1       2      2        2        2       2        2       2
#> [2,]       1       1       2      2        2        2       2        2       2
#> [3,]       1       1       2      2        2        2       2        2       2
#> [4,]       5       1       2      2        2        2       2        2       2
#> [5,]       5       4       2      2        2        2       2        2       2
#> [6,]       5       4       2      2        2        2       2        2       2
#>      POSLABLY POSFACT COMPCODE OCCUCODE KINDFACT TENURE FTPTIME ADDJOB JOBFIND
#> [1,]        7       1       95      300        8      1       1      2       2
#> [2,]        7       1       67      552        8      1       2      2       2
#> [3,]        7       1       90      219        8      1       1      2       2
#> [4,]        7       1       65      481        8      1       2      2       2
#> [5,]        2       7        0        0       97      7       7      7       7
#> [6,]        2       7        0        0       97      7       7      7       7
#>      WEIGHT INCOME MONEY ASSETS DEBTS
#> [1,]  118.3 970000 74231  61340  1950
#> [2,]  118.3 393700 74231  90480  6750
#> [3,]  118.3 152200 74231  96600  6970
#> [4,]  118.3 216100 74231  40870  5950
#> [5,]  118.3 566900 74231  66460  4610
#> [6,]  118.3 776700 74231  46870  3900