Plots for localSuppression objects

This function creates barplots to display the number of suppressed values in categorical key variables to achieve k-anonymity.

# S3 method for localSuppression
plot(x, ...)

Arguments

x

object of derived from localSuppression()

...

Additional arguments, currently available are:

"showDetails": logical, if set, a plot of suppressions by strata is shown (if possible)

Value

a ggplot plot object

See also

Author

Bernhard Meindl, Matthias Templ

Examples

data(francdat)
#l1 <- localSuppression(obj = francdat, k = 2, keyVars = c(2, 4:6))
#l1
#plot(l1)
#
### with details of suppression by strata
#data(testdata2)
#testdata2$ageG <- cut(testdata2$age, 5, labels = paste0("AG",1:5))
#keyVars <- c("urbrur", "roof", "walls", "water", "electcon", "relat", "sex")
#strataVars <- "ageG"
#inp <- testdata2[, c(keyVars, strataVars)]
#ls <- localSuppression(inp, keyVars = 1:7, strataVars = 8)
#print(ls)
#plot(ls)
#plot(ls, showDetails = TRUE)