Post Randomization
Source: R/pram.R

To be used on categorical data stored as factors. The algorithm randomly changes the values of variables in selected records (usually the risky ones) according to an invariant probability transition matrix or a custom-defined transition matrix.

pram(obj, variables = NULL, strata_variables = NULL, pd = 0.8,
  alpha = 0.5)

Arguments

obj
Input data. Allowed input data are objects of class data.frame, factor or sdcMicroObj-class.

variables
Names of variables in 'obj' on which post-randomization should be applied. If obj is a factor, this argument is ignored. Please note that pram can only be applied to factor-variables.

strata_variables
Names of variables for stratification (will be set automatically for an object of class sdcMicroObj-class. One can also specify an integer vector or factor that specifies that desired groups. This vector must match the dimension of the input data set, however. For a possible use case, have a look at the examples.

pd
minimum diagonal entries for the generated transition matrix P. Either a vector of length 1 (which is recycled) or a vector of the same length as the number of variables that should be postrandomized. It is also possible to set pd to a numeric matrix. This matrix will be used directly as the transition matrix. The matrix must be constructed as follows:

the matrix must be a square matrix

the rownames and colnames of the matrix must match the levels (in the same order) of the factor-variable that should be postrandomized.

the rowSums and colSums of the matrix need to equal 1

It is also possible to combine the different ways. For details have a look at the examples.

alpha
amount of perturbation for the invariant Pram method. This is a numeric vector of length 1 (that will be recycled if necessary) or a vector of the same length as the number of variables. If one specified as transition matrix directly, alpha is ignored.

...
further input, currently ignored.

Value

a modified sdcMicroObj-class object or a new object containing original and post-randomized variables (with suffix "_pram").

Note

Deprecated method 'pram_strata'is no longer available in sdcMicro > 4.5.0

References

http://www.gnu.org/software/glpk

Kowarik, A. and Templ, M. and Meindl, B. and Fonteneau, F. and Prantner, B.: Testing of IHSN Cpp Code and Inclusion of New Methods into sdcMicro, in: Lecture Notes in Computer Science, J. Domingo-Ferrer, I. Tinnirello (editors.); Springer, Berlin, 2012, ISBN: 978-3-642-33626-3, pp. 63-77. doi: 10.1007/978-3-642-33627-0_6

Templ, M. and Kowarik, A. and Meindl, B. Statistical Disclosure Control for Micro-Data Using the R Package sdcMicro. Journal of Statistical Software, 67 (4), 1--36, 2015. doi: 10.18637/jss.v067.i04

Templ, M. Statistical Disclosure Control for Microdata: Methods and Applications in R. Springer International Publishing, 287 pages, 2017. ISBN 978-3-319-50272-4. doi: 10.1007/978-3-319-50272-4

Examples

data(testdata)

## application on a factor-variable
res <- pram(as.factor(testdata$roof))
print(res)

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> x != x_pram : 224 (4.89%)

summary(res)

#> Variable:  x
#>  ----------------------
#> Frequencies in original and perturbed data:
#>                                 x   2    4  5  6  9 NA
#> 1:           Original Frequencies 814 3697 19 34 16  0
#> 2: Frequencies after Perturbation 801 3729 15 23 12  0
#> 
#> Transitions:
#>     transition Frequency
#>  1:    1 --> 1       721
#>  2:    1 --> 2        91
#>  3:    1 --> 3         1
#>  4:    1 --> 4         1
#>  5:    2 --> 1        78
#>  6:    2 --> 2      3603
#>  7:    2 --> 3         5
#>  8:    2 --> 4         5
#>  9:    2 --> 5         6
#> 10:    3 --> 2        10
#> 11:    3 --> 3         9
#> 12:    4 --> 2        17
#> 13:    4 --> 4        17
#> 14:    5 --> 1         2
#> 15:    5 --> 2         8
#> 16:    5 --> 5         6
#> 

## application on a data.frame
## pram can only be applied to factors, thus we have to recode
## to factors before the method can be applied
testdata$roof <- factor(testdata$roof)
testdata$walls <- factor(testdata$walls)
testdata$water <- factor(testdata$water)

## pram() is applied within subgroups defined by
## variables "urbrur" and "sex"
res <- pram(testdata, variables="roof",
  strata_variables=c("urbrur","sex"))
print(res)

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> roof != roof_pram : 227 (4.96%)

summary(res)

#> Variable:  roof
#>  ----------------------
#> Frequencies in original and perturbed data:
#>                              roof   2    4  5  6  9 NA
#> 1:           Original Frequencies 814 3697 19 34 16  0
#> 2: Frequencies after Perturbation 818 3704 13 37  8  0
#> 
#> Transitions:
#>     transition Frequency
#>  1:    1 --> 1       722
#>  2:    1 --> 2        85
#>  3:    1 --> 3         1
#>  4:    1 --> 4         5
#>  5:    1 --> 5         1
#>  6:    2 --> 1        91
#>  7:    2 --> 2      3593
#>  8:    2 --> 3         1
#>  9:    2 --> 4         9
#> 10:    2 --> 5         3
#> 11:    3 --> 2         8
#> 12:    3 --> 3        11
#> 13:    4 --> 1         3
#> 14:    4 --> 2         8
#> 15:    4 --> 4        23
#> 16:    5 --> 1         2
#> 17:    5 --> 2        10
#> 18:    5 --> 5         4
#> 

## default parameters (pd=0.8 and alpha=0.5) for the generation
## of the invariant transition matrix will be used for all variables
res1 <- pram(testdata, variables=c("roof","walls","water"))
print(res1)

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> roof != roof_pram : 108 (2.36%)
#> walls != walls_pram : 216 (4.72%)
#> water != water_pram : 268 (5.85%)

## specific parameters for each variable
res2 <- pram(testdata,variables=c("roof","walls","water"),
  pd=c(0.95,0.8,0.9), alpha=0.5)
print(res2)

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> roof != roof_pram : 101 (2.21%)
#> walls != walls_pram : 364 (7.95%)
#> water != water_pram : 265 (5.79%)

## detailed information on pram-parameters (such as the transition matrix 'Rs')
## is stored in the output, eg. for variable 'roof'
attr(res2, "pram_params")$roof

#> $Rs
#>             2          4           5           6           9
#> 2 0.954567185 0.03951272 0.001780602 0.002544967 0.001594524
#> 4 0.008699853 0.98634404 0.001481717 0.002169618 0.001304777
#> 5 0.076284748 0.28831094 0.631545979 0.001746827 0.002111505
#> 6 0.060929496 0.23591404 0.000976168 0.700947786 0.001232506
#> 9 0.081121412 0.30148510 0.002507412 0.002619076 0.612267001
#> 
#> $pd
#> [1] 0.95
#> 
#> $alpha
#> [1] 0.5
#> 

## we can also specify a custom transition-matrix directly
# for variable roof; matrix must have rownames and colnames that match
# the levels of the variable that should be post-randomized
# rowSums() and colSums() must equal 1 too!
mat <- diag(length(levels(testdata$roof)))
rownames(mat) <- colnames(mat) <- levels(testdata$roof)
res3 <- pram(testdata,variables="roof", pd=mat)
print(res3) # of course, nothing has changed!

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> roof != roof_pram : 0 (0%)

## it is possible use a transistion matrix for a variable and use the 'traditional' way
## of specifying a number for the minimal diagonal entries of the transision matrix
## for other variables. In this case we must supply \code{pd} as list.
res4 <- pram(testdata,variables=c("roof","walls"), pd=list(mat,0.5), alpha=c(NA, 0.5))
print(res4)

#> Number of changed observations: 
#> - - - - - - - - - - - 
#> roof != roof_pram : 0 (0%)
#> walls != walls_pram : 456 (9.96%)

summary(res4)

#> Variable:  roof
#>  ----------------------
#> Frequencies in original and perturbed data:
#>                              roof   2    4  5  6  9 NA
#> 1:           Original Frequencies 814 3697 19 34 16  0
#> 2: Frequencies after Perturbation 814 3697 19 34 16  0
#> 
#> Transitions:
#>    transition Frequency
#> 1:    1 --> 1       814
#> 2:    2 --> 2      3697
#> 3:    3 --> 3        19
#> 4:    4 --> 4        34
#> 5:    5 --> 5        16
#> 
#> Variable:  walls
#>  ----------------------
#> Frequencies in original and perturbed data:
#>                             walls    2    3  9 NA
#> 1:           Original Frequencies 1203 3327 50  0
#> 2: Frequencies after Perturbation 1197 3353 30  0
#> 
#> Transitions:
#>    transition Frequency
#> 1:    1 --> 1       982
#> 2:    1 --> 2       220
#> 3:    1 --> 3         1
#> 4:    2 --> 1       206
#> 5:    2 --> 2      3117
#> 6:    2 --> 3         4
#> 7:    3 --> 1         9
#> 8:    3 --> 2        16
#> 9:    3 --> 3        25
#> 

attr(res4, "pram_params")

#> $roof
#> $roof$Rs
#>   2 4 5 6 9
#> 2 1 0 0 0 0
#> 4 0 1 0 0 0
#> 5 0 0 1 0 0
#> 6 0 0 0 1 0
#> 9 0 0 0 0 1
#> 
#> $roof$pd
#>   2 4 5 6 9
#> 2 1 0 0 0 0
#> 4 0 1 0 0 0
#> 5 0 0 1 0 0
#> 6 0 0 0 1 0
#> 9 0 0 0 0 1
#> 
#> $roof$alpha
#> [1] NA
#> 
#> 
#> $walls
#> $walls$Rs
#>            2         3           9
#> 2 0.80649995 0.1878785 0.005621558
#> 3 0.06793442 0.9271329 0.004932626
#> 9 0.13525469 0.3282169 0.536528377
#> 
#> $walls$pd
#> [1] 0.5
#> 
#> $walls$alpha
#> [1] 0.5
#> 
#> 

## application to objects of class sdcMicro with default parameters
data(testdata2)
testdata2$urbrur <- factor(testdata2$urbrur)
sdc <- createSdcObj(testdata2,
  keyVars=c('roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
sdc <- pram(sdc, variables=c("urbrur"))
print(sdc, type="pram")

#> Post-Randomization (PRAM):
#> Variable: urbrur 
#> --> final Transition-Matrix:
#>            1         2
#> 1 0.83503881 0.1649612
#> 2 0.06748412 0.9325159
#> 
#> Changed observations:
#>   variable nrChanges percChanges
#> 1   urbrur        12        12.9
#> ---------------------------------------------------------------------- 
#> 

## this is equal to the previous application. If argument 'variables' is NULL,
## all variables from slot 'pramVars' will be used if possible.
sdc <- createSdcObj(testdata2,
  keyVars=c('roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight',
  pramVars="urbrur")
sdc <- pram(sdc)
print(sdc, type="pram")

#> Post-Randomization (PRAM):
#> Variable: urbrur 
#> --> final Transition-Matrix:
#>            1         2
#> 1 0.89575182 0.1042482
#> 2 0.04264698 0.9573530
#> 
#> Changed observations:
#>   variable nrChanges percChanges
#> 1   urbrur        10       10.75
#> ---------------------------------------------------------------------- 
#> 

## we can specify transition matrices for sdcMicroObj-objects too
testdata2$roof <- factor(testdata2$roof)
sdc <- createSdcObj(testdata2,
  keyVars=c('roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
mat <- diag(length(levels(testdata2$roof)))
rownames(mat) <- colnames(mat) <- levels(testdata2$roof)
mat[1,] <- c(0.9,0,0,0.05,0.05)
sdc <- pram(sdc, variables="roof", pd=mat)

#> Warning: If pram is applied on key variables, the k-anonymity and risk assessment are not useful anymore.

print(sdc, type="pram")

#> Post-Randomization (PRAM):
#> Variable: roof 
#> --> final Transition-Matrix:
#>     2 4 5    6    9
#> 2 0.9 0 0 0.05 0.05
#> 4 0.0 1 0 0.00 0.00
#> 5 0.0 0 1 0.00 0.00
#> 6 0.0 0 0 1.00 0.00
#> 9 0.0 0 0 0.00 1.00
#> 
#> Changed observations:
#>   variable nrChanges percChanges
#> 1     roof         2        2.15
#> ---------------------------------------------------------------------- 
#> 

# we can also have a look at the transitions
get.sdcMicroObj(sdc, "pram")$transitions

#> $roof
#>    transition Frequency
#> 1:    1 --> 1        25
#> 2:    1 --> 4         2
#> 3:    2 --> 2        46
#> 4:    3 --> 3         6
#> 5:    4 --> 4         8
#> 6:    5 --> 5         6
#>