Print method for objects from class micro
Source: R/microaggregation.R

printing an object of class micro

# S3 method for micro
print(x, ...)

Arguments

x
object from class micro

…
Additional arguments passed through.

Value

information about method and aggregation level from objects of class micro.

See also

microaggregation

Examples

data(free1)
free1 <- as.data.frame(free1)
m1 <- microaggregation(free1[, 31:34], method='onedims', aggr=3)
m1

#> 
#>  Object created with method onedims and aggregation level 3
#>  -------------------------
#> x ... original values 
#>      INCOME           MONEY           ASSETS          DEBTS     
#>  Min.   :100100   Min.   : 5532   Min.   :10000   Min.   :1000  
#>  1st Qu.:329075   1st Qu.:32311   1st Qu.:31695   1st Qu.:3340  
#>  Median :549400   Median :53521   Median :53975   Median :5690  
#>  Mean   :547254   Mean   :52610   Mean   :54300   Mean   :5587  
#>  3rd Qu.:764875   3rd Qu.:74231   3rd Qu.:76298   3rd Qu.:7840  
#>  Max.   :999400   Max.   :95431   Max.   :99960   Max.   :9990  
#> 
#>  -------------------------
#> mx ... microaggregated values
#>      INCOME           MONEY           ASSETS          DEBTS     
#>  Min.   :100300   Min.   : 5532   Min.   :10020   Min.   :1000  
#>  1st Qu.:329167   1st Qu.:32311   1st Qu.:31707   1st Qu.:3340  
#>  Median :549400   Median :53521   Median :53953   Median :5690  
#>  Mean   :547254   Mean   :52610   Mean   :54300   Mean   :5587  
#>  3rd Qu.:764792   3rd Qu.:74231   3rd Qu.:76299   3rd Qu.:7837  
#>  Max.   :999075   Max.   :95431   Max.   :99848   Max.   :9990  
#> 
#>  -------------------------
#> Try names(your object from class micro) for more details