Generate an Html-report from an sdcMicroObj

Summary statistics of the original and the perturbed data set

report(
  obj,
  outdir = tempdir(),
  filename = "SDC-Report",
  title = "SDC-Report",
  internal = FALSE,
  verbose = FALSE
)

Arguments

obj

an object of class sdcMicroObj-class or reportObj

outdir

output folder

filename

output filename

title

Title for the report

internal

TRUE/FALSE, if TRUE a detailed internal report is produced, else a non-disclosive overview

verbose

TRUE/FALSE, if TRUE, some additional information is printed.

Details

The application of this function provides you with a html-report for your sdcMicro object that contains useful summaries about the anonymization process.

Author

Matthias Templ, Bernhard Meindl

Examples

# \donttest{
data(testdata2)
sdc <- createSdcObj(
  dat = testdata2,
  keyVars = c("urbrur", "roof", "walls", "water", "electcon", "relat", "sex"),
  numVars = c("expend", "income", "savings"),
  w = "sampling_weight"
)
report(sdc)
# }