RMD based disclosure risk

Distance-based disclosure risk estimation via robust Mahalanobis Distances.

dRiskRMD(obj, ...)

Arguments

obj

an sdcMicroObj-class-object or a data.frame

...

see possible arguments below

xm masked data

kweight for adjusting the influence of the robust Mahalanobis distances, i.e. to increase or decrease each of the disclosure risk intervals.

k2parameter for method RMDID2 to choose a small interval around each masked observation.

Value

The disclosure risk or the modified sdcMicroObj-class

risk1

percentage of sensitive observations according to method RMDID1.

risk2

standardized version of risk1

wrisk1

amount of sensitive observations according to RMDID1 weighted by their corresponding robust Mahalanobis distances.

wrisk2

RMDID2 measure

indexRisk1

index of observations with high risk according to risk1 measure

indexRisk2

index of observations with high risk according to wrisk2 measure

Details

This method is an extension of method SDID because it accounts for the “outlyingness” of each observations. This is a quite natural approach since outliers do have a higher risk of re-identification and therefore these outliers should have larger disclosure risk intervals as observations in the center of the data cloud.

The algorithm works as follows:

1. Robust Mahalanobis distances are estimated in order to get a robust multivariate distance for each observation.

2. Intervals are estimated for each observation around every data point of the original data points where the length of the interval is defined/weighted by the squared robust Mahalanobis distance and the parameter $k$. The higher the RMD of an observation the larger the interval.

3. Check if the corresponding masked values fall into the intervals around the original values or not. If the value of the corresponding observation is within such an interval the whole observation is considered unsafe. So, we get a whole vector indicating which observation is save or not, and we are finished already when using method RMDID1).

4. For method RMDID1w: we return the weighted (via RMD) vector of disclosure risk.

5. For method RMDID2: whenever an observation is considered unsafe it is checked if $m$ other observations from the masked data are very close (defined by a parameter $k2$ for the length of the intervals as for SDID or RSDID) to such an unsafe observation from the masked data, using Euclidean distances. If more than $m$ points are in such a small interval, we conclude that this observation is ``save''.

References

Templ, M. and Meindl, B., Robust Statistics Meets SDC: New Disclosure Risk Measures for Continuous Microdata Masking, Lecture Notes in Computer Science, Privacy in Statistical Databases, vol. 5262, pp. 113-126, 2008.

Templ, M. New Developments in Statistical Disclosure Control and Imputation: Robust Statistics Applied to Official Statistics, Suedwestdeutscher Verlag fuer Hochschulschriften, 2009, ISBN: 3838108280, 264 pages.

See also

Author

Matthias Templ

Examples

data(Tarragona)
x <- Tarragona[, 5:7]
y <- addNoise(x)$xm
dRiskRMD(x, xm=y)
#> $risk1
#> [1] 0.008393285
#> 
#> $risk2
#> [1] 0.008393285
#> 
#> $wrisk1
#> [1] 0.01159607
#> 
#> $wrisk2
#> [1] 0.01159607
#> 
#> $indexRisk1
#>  33 440 526 551 616 667 673 
#>  33 440 526 551 616 667 673 
#> 
#> $indexRisk2
#> [1]  33 440 526 551 616 667 673
#> 
#> $riskvec1
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.19044119 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [61] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [79] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [85] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [91] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [97] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [103] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [109] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [115] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [121] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [127] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [133] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [139] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [145] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [151] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [157] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [163] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [169] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [175] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [181] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [187] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [193] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [199] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [205] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [211] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [217] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [223] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [229] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [235] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [241] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [247] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [253] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [259] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [265] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [271] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [277] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [283] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [289] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [295] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [301] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [307] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [313] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [319] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [325] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [331] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [337] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [343] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [349] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [355] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [361] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [367] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [373] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [379] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [385] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [391] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [397] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [403] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [409] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [415] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [421] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [427] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [433] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [439] 0.00000000 0.14925774 0.00000000 0.00000000 0.00000000 0.00000000
#> [445] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [451] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [457] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [463] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [469] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [475] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [481] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [487] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [493] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [499] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [505] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [511] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [517] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [523] 0.00000000 0.00000000 0.00000000 0.19922235 0.00000000 0.00000000
#> [529] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [535] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [541] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [547] 0.00000000 0.00000000 0.00000000 0.00000000 0.15993346 0.00000000
#> [553] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [559] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [565] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [571] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [577] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [583] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [589] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [595] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [601] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [607] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [613] 0.00000000 0.00000000 0.00000000 0.09852538 0.00000000 0.00000000
#> [619] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [625] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [631] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [637] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [643] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [649] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [655] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [661] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [667] 0.34959297 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [673] 0.23461621 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [679] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [685] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [691] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [697] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [703] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [709] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [715] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [721] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [727] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [733] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [739] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [745] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [751] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [757] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [763] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [769] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [775] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [781] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [787] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [793] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [799] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [805] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [811] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [817] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [823] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [829] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
#> $riskvec2
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.19044119 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [61] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [79] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [85] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [91] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [97] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [103] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [109] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [115] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [121] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [127] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [133] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [139] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [145] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [151] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [157] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [163] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [169] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [175] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [181] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [187] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [193] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [199] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [205] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [211] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [217] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [223] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [229] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [235] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [241] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [247] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [253] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [259] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [265] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [271] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [277] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [283] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [289] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [295] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [301] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [307] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [313] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [319] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [325] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [331] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [337] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [343] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [349] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [355] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [361] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [367] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [373] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [379] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [385] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [391] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [397] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [403] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [409] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [415] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [421] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [427] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [433] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [439] 0.00000000 0.14925774 0.00000000 0.00000000 0.00000000 0.00000000
#> [445] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [451] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [457] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [463] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [469] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [475] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [481] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [487] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [493] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [499] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [505] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [511] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [517] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [523] 0.00000000 0.00000000 0.00000000 0.19922235 0.00000000 0.00000000
#> [529] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [535] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [541] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [547] 0.00000000 0.00000000 0.00000000 0.00000000 0.15993346 0.00000000
#> [553] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [559] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [565] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [571] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [577] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [583] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [589] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [595] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [601] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [607] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [613] 0.00000000 0.00000000 0.00000000 0.09852538 0.00000000 0.00000000
#> [619] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [625] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [631] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [637] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [643] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [649] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [655] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [661] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [667] 0.34959297 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [673] 0.23461621 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [679] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [685] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [691] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [697] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [703] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [709] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [715] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [721] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [727] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [733] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [739] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [745] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [751] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [757] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [763] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [769] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [775] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [781] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [787] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [793] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [799] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [805] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [811] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [817] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [823] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [829] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
dRisk(x, xm=y)
#> [1] 0

data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
## this is already made internally:
## sdc <- dRiskRMD(sdc)
## and already stored in sdc