Data-Utility measures

dUtility() allows to compute different measures of data-utility based on various distances using original and perturbed variables.

dUtility(obj, ...)

Arguments

obj

original data or object of class sdcMicroObj

...

see arguments below

xm: perturbed data

method: method IL1, IL1s or eigen. More methods are implemented in summary.micro()

Value

data utility or modified entry for data utility the sdcMicroObj.

Details

The standardised distances of the perturbed data values to the original ones are measured. The following measures are available:

"IL1: sum of absolute distances between original and perturbed variables scaled by absolute values of the original variables

"IL1s: measures the absolute distances between original and perturbed ones, scaled by the standard deviation of original variables times the square root of 2.

"eigen; compares the eigenvalues of original and perturbed data

"robeigen; compares robust eigenvalues of original and perturbed data

References

for IL1 and IL1s: see Mateo-Sanz, Sebe, Domingo-Ferrer. Outlier Protection in Continuous Microdata Masking. International Workshop on Privacy in Statistical Databases. PSD 2004: Privacy in Statistical Databases pp 201-215.

Templ, M. and Meindl, B., Robust Statistics Meets SDC: New Disclosure Risk Measures for Continuous Microdata Masking, Lecture Notes in Computer Science, Privacy in Statistical Databases, vol. 5262, pp. 113-126, 2008.

See also

Author

Matthias Templ