Data-Utility measures

dUtility() allows to compute different measures of data-utility based on various distances using original and perturbed variables.

dUtility(obj, ...)

Arguments

obj

original data or object of class sdcMicroObj

...

see arguments below

xm: perturbed data

method: method IL1, IL1s or eigen. More methods are implemented in summary.micro()

Value

data utility or modified entry for data utility the sdcMicroObj.

Details

The standardised distances of the perturbed data values to the original ones are measured. The following measures are available:

"IL1: sum of absolute distances between original and perturbed variables scaled by absolute values of the original variables

"IL1s: measures the absolute distances between original and perturbed ones, scaled by the standard deviation of original variables times the square root of 2.

"eigen; compares the eigenvalues of original and perturbed data

"robeigen; compares robust eigenvalues of original and perturbed data

References

for IL1 and IL1s: see Mateo-Sanz, Sebe, Domingo-Ferrer. Outlier Protection in Continuous Microdata Masking. International Workshop on Privacy in Statistical Databases. PSD 2004: Privacy in Statistical Databases pp 201-215.

Templ, M. and Meindl, B., Robust Statistics Meets SDC: New Disclosure Risk Measures for Continuous Microdata Masking, Lecture Notes in Computer Science, Privacy in Statistical Databases, vol. 5262, pp. 113-126, 2008.

See also

Author

Matthias Templ

Examples

data(free1)
free1 <- as.data.frame(free1)
m1 <- microaggregation(free1[, 31:34], method="onedims", aggr=3)
m2 <- microaggregation(free1[, 31:34], method="pca", aggr=3)
dRisk(obj=free1[, 31:34], xm=m1$mx)
#> [1] 0.9955
dRisk(obj=free1[, 31:34], xm=m2$mx)
#> [1] 0
dUtility(obj=free1[, 31:34], xm=m1$mx)
#> [1] 7.971673
dUtility(obj=free1[, 31:34], xm=m2$mx)
#> [1] 6335.63
data(Tarragona)
x <- Tarragona[, 5:7]
y <- addNoise(x)$xm
dRiskRMD(x, xm=y)
#> $risk1
#> [1] 0.007194245
#> 
#> $risk2
#> [1] 0.007194245
#> 
#> $wrisk1
#> [1] 0.04053893
#> 
#> $wrisk2
#> [1] 0.04053893
#> 
#> $indexRisk1
#> 199 352 677 778 795 804 
#> 199 352 677 778 795 804 
#> 
#> $indexRisk2
#> [1] 199 352 677 778 795 804
#> 
#> $riskvec1
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [61] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [79] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [85] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [91] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [97] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [103] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [109] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [115] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [121] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [127] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [133] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [139] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [145] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [151] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [157] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [163] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [169] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [175] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [181] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [187] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [193] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [199] 0.16505390 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [205] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [211] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [217] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [223] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [229] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [235] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [241] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [247] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [253] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [259] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [265] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [271] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [277] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [283] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [289] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [295] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [301] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [307] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [313] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [319] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [325] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [331] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [337] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [343] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [349] 0.00000000 0.00000000 0.00000000 0.24849691 0.00000000 0.00000000
#> [355] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [361] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [367] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [373] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [379] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [385] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [391] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [397] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [403] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [409] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [415] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [421] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [427] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [433] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [439] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [445] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [451] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [457] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [463] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [469] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [475] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [481] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [487] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [493] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [499] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [505] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [511] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [517] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [523] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [529] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [535] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [541] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [547] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [553] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [559] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [565] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [571] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [577] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [583] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [589] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [595] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [601] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [607] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [613] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [619] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [625] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [631] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [637] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [643] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [649] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [655] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [661] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [667] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [673] 0.00000000 0.00000000 0.00000000 0.00000000 0.08030945 0.00000000
#> [679] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [685] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [691] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [697] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [703] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [709] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [715] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [721] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [727] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [733] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [739] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [745] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [751] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [757] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [763] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [769] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [775] 0.00000000 0.00000000 0.00000000 0.60568232 0.00000000 0.00000000
#> [781] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [787] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [793] 0.00000000 0.00000000 2.83257590 0.00000000 0.00000000 0.00000000
#> [799] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 1.70279279
#> [805] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [811] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [817] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [823] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [829] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
#> $riskvec2
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [61] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [79] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [85] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [91] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [97] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [103] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [109] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [115] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [121] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [127] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [133] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [139] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [145] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [151] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [157] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [163] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [169] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [175] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [181] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [187] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [193] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [199] 0.16505390 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [205] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [211] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [217] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [223] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [229] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [235] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [241] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [247] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [253] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [259] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [265] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [271] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [277] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [283] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [289] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [295] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [301] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [307] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [313] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [319] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [325] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [331] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [337] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [343] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [349] 0.00000000 0.00000000 0.00000000 0.24849691 0.00000000 0.00000000
#> [355] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [361] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [367] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [373] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [379] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [385] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [391] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [397] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [403] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [409] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [415] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [421] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [427] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [433] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [439] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [445] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [451] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [457] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [463] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [469] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [475] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [481] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [487] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [493] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [499] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [505] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [511] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [517] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [523] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [529] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [535] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [541] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [547] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [553] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [559] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [565] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [571] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [577] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [583] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [589] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [595] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [601] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [607] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [613] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [619] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [625] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [631] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [637] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [643] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [649] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [655] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [661] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [667] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [673] 0.00000000 0.00000000 0.00000000 0.00000000 0.08030945 0.00000000
#> [679] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [685] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [691] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [697] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [703] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [709] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [715] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [721] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [727] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [733] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [739] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [745] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [751] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [757] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [763] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [769] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [775] 0.00000000 0.00000000 0.00000000 0.60568232 0.00000000 0.00000000
#> [781] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [787] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [793] 0.00000000 0.00000000 2.83257590 0.00000000 0.00000000 0.00000000
#> [799] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 1.70279279
#> [805] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [811] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [817] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [823] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [829] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
dRisk(x, xm=y)
#> [1] 0
dUtility(x, xm = y, method = "IL1")
#> [1] 3571151
dUtility(x, xm = y, method = "IL1s")
#> [1] 1911.932
dUtility(x, xm = y, method = "eigen")
#> [1] 3.260089
dUtility(x, xm = y, method = "robeigen")
#> [1] 167.286

## for objects of class sdcMicro:
data(testdata2)
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')
## this is already made internally:
## sdc <- dUtility(sdc)
## and already stored in sdc