data from the casc project

Small synthetic data from Capobianchi, Polettini, Lucarelli

Format

A data frame with 8 observations on the following 8 variables.

Num1

a numeric vector

Key1

Key variable 1. A numeric vector

Num2

a numeric vector

Key2

Key variable 2. A numeric vector

Key3

Key variable 3. A numeric vector

Key4

Key variable 4. A numeric vector

Num3

a numeric vector

w

The weight vector. A numeric vector

Source

https://research.cbs.nl/casc/deliv/12d1.pdf

Details

This data set is very similar to that one which are used by the authors of the paper given below. We need this data set only for demonstration effect, i.e. that the package provides the same results as their software.

Examples

data(francdat)
francdat
#>   Num1 Key1 Num2 Key2 Key3 Key4 Num3     w
#> 1 0.30    1 0.40    2    5    1    4  18.0
#> 2 0.12    1 0.22    2    1    1   22  45.5
#> 3 0.18    1 0.80    2    1    1    8  39.0
#> 4 1.90    3 9.00    3    1    5   91  17.0
#> 5 1.00    4 1.30    3    1    4   13 541.0
#> 6 1.00    4 1.40    3    1    1   14   8.0
#> 7 0.10    6 0.01    2    1    5    1   5.0
#> 8 0.15    1 0.50    2    5    1    5  92.0