Individual Risk computation
Source: R/indivRisk.R

Estimation of the risk for each observation. After the risk is computed one can use e.g. the function localSuppr() for the protection of values of high risk. Further details can be found at the link given below.

indivRisk(x, method = "approx", qual = 1, survey = TRUE)

Arguments

x
object from class freqCalc

method
approx (default) or exact

qual
final correction factor

survey
TRUE, if we have survey data and FALSE if we deal with a population.

Value

rk: base individual risk

method: method

qual: final correction factor

fk: frequency count

knames: colnames of the key variables

Details

S4 class sdcMicro objects are only supported by function measure_risk that also estimates the individual risk with the same method.

Note

The base individual risk method was developed by Benedetti, Capobianchi and Franconi

References

Templ, M. and Kowarik, A. and Meindl, B. Statistical Disclosure Control for Micro-Data Using the R Package sdcMicro. Journal of Statistical Software, 67 (4), 1--36, 2015. doi: 10.18637/jss.v067.i04

Franconi, L. and Polettini, S. (2004) Individual risk estimation in mu-Argus: a review. Privacy in Statistical Databases, Lecture Notes in Computer Science, 262--272. Springer

Machanavajjhala, A. and Kifer, D. and Gehrke, J. and Venkitasubramaniam, M. (2007) l-Diversity: Privacy Beyond k-Anonymity. ACM Trans. Knowl. Discov. Data, 1(1)

additionally, have a look at the vignettes of sdcMicro for further reading.

See also

measure_risk, freqCalc

Examples

## example from Capobianchi, Polettini and Lucarelli:
data(francdat)
f <- freqCalc(francdat, keyVars=c(2,4,5,6),w=8)
f

#> 
#>  --------------------------
#> 4 obs. violate 2-anonymity 
#> 8 obs. violate 3-anonymity 
#>  --------------------------

f$fk

#> [1] 2 2 2 1 1 1 1 2

f$Fk

#> [1] 110.0  84.5  84.5  17.0 541.0   8.0   5.0 110.0

## individual risk calculation:
indivf <- indivRisk(f)
indivf$rk

#> [1] 0.01714426 0.02204233 0.02204233 0.17707583 0.01165448 0.29706308 0.40235948
#> [8] 0.01714426