kAnon_violations

returns the number of observations violating k-anonymity.

kAnon_violations(object, weighted, k)

# S4 method for sdcMicroObj,logical,numeric
kAnon_violations(object, weighted, k)

Arguments

object

a sdcMicroObj-class object

weighted

TRUE or FALSE defining if sampling weights should be taken into account

k

a positive number defining parameter k

Value

the number of records that are violating k-anonymity based on unweighted sample data only (in case parameter weighted is FALSE) or computing the number of observations that are estimated to violate k-anonymity in the population in case parameter weighted equals TRUE.