riskyCells
Source: R/riskyCells.R

Allows to compute risky (unweighted) combinations of key variables either up to a specified dimension or using identification level. This mimics the approach taken in mu-argus.

riskyCells(obj, useIdentificationLevel = FALSE, threshold, ...)

Arguments

obj
a data.frame, data.table or an object of class sdcMicroObj-class

useIdentificationLevel
(logical) specifies if tabulation should be done up to a specific dimension (useIdentificationLevel=FALSE using argument maxDim) or taking identification levels (useIdentificationLevel=FALSE using argument level) into account.

threshold
a numeric vector specifiying the thresholds at which cells are considered to be unsafe. In case a tabulation is done up to a specific level (useIdentificationLevel=FALSE), the thresholds may be specified differently for each dimension. In the other case, the same threshold is used for all tables.

...
see possible arguments below

keyVars: index or variable-names within obj that should be used for tabulation. In case obj is of class sdcMicroObj-class, this argument is not used and the pre-defined key-variables are used.

level: in case useIdentificationLevel=TRUE, this numeric vector specifies the importance of the key variables. The construction of output tables follows the implementation in mu-argus, see e.g http://neon.vb.cbs.nl/casc/Software/MUmanual5.1.pdf. The length of this numeric vector must match the number of key variables.

maxDim: in case useIdentificationLevel=FALSE, this number specifies maximal number of variables to tablulate.

Value

a data.table showing the number of unsafe cells, thresholds for any combination of the key variables. If the input was a sdcMicroObj-class object and some modifications have been already applied to the categorical key variables, the resulting output contains the number of unsafe cells both for the original and the modified data.

Examples

## data.frame method / all combinations up to maxDim
riskyCells(testdata2, keyVars=c(1:5), threshold=c(50,25,10,5),
  useIdentificationLevel=FALSE, maxDim=4)

#>         dim1     dim2     dim3     dim4 threshold unsafe_cells
#>  1:   urbrur       NA       NA       NA        50            1
#>  2:     roof       NA       NA       NA        50            5
#>  3:    walls       NA       NA       NA        50            2
#>  4:    water       NA       NA       NA        50            8
#>  5: electcon       NA       NA       NA        50            1
#>  6:   urbrur     roof       NA       NA        25            8
#>  7:   urbrur    walls       NA       NA        25            5
#>  8:   urbrur    water       NA       NA        25           12
#>  9:   urbrur electcon       NA       NA        25            2
#> 10:     roof    walls       NA       NA        25            9
#> 11:     roof    water       NA       NA        25           24
#> 12:     roof electcon       NA       NA        25            8
#> 13:    walls    water       NA       NA        25           17
#> 14:    walls electcon       NA       NA        25            4
#> 15:    water electcon       NA       NA        25           13
#> 16:   urbrur     roof    walls       NA        10           12
#> 17:   urbrur     roof    water       NA        10           32
#> 18:   urbrur     roof electcon       NA        10           10
#> 19:   urbrur    walls    water       NA        10           24
#> 20:   urbrur    walls electcon       NA        10            6
#> 21:   urbrur    water electcon       NA        10           19
#> 22:     roof    walls    water       NA        10           36
#> 23:     roof    walls electcon       NA        10           14
#> 24:     roof    water electcon       NA        10           31
#> 25:    walls    water electcon       NA        10           26
#> 26:   urbrur     roof    walls    water         5           47
#> 27:   urbrur     roof    walls electcon         5           18
#> 28:   urbrur     roof    water electcon         5           35
#> 29:   urbrur    walls    water electcon         5           32
#> 30:     roof    walls    water electcon         5           44
#>         dim1     dim2     dim3     dim4 threshold unsafe_cells

riskyCells(testdata2, keyVars=c(1:5), threshold=10,
  useIdentificationLevel=FALSE, maxDim=3)

#>         dim1     dim2     dim3 threshold unsafe_cells
#>  1:   urbrur       NA       NA        10            0
#>  2:     roof       NA       NA        10            3
#>  3:    walls       NA       NA        10            1
#>  4:    water       NA       NA        10            4
#>  5: electcon       NA       NA        10            0
#>  6:   urbrur     roof       NA        10            6
#>  7:   urbrur    walls       NA        10            2
#>  8:   urbrur    water       NA        10            9
#>  9:   urbrur electcon       NA        10            1
#> 10:     roof    walls       NA        10            6
#> 11:     roof    water       NA        10           24
#> 12:     roof electcon       NA        10            6
#> 13:    walls    water       NA        10           16
#> 14:    walls electcon       NA        10            3
#> 15:    water electcon       NA        10            9
#> 16:   urbrur     roof    walls        10           12
#> 17:   urbrur     roof    water        10           32
#> 18:   urbrur     roof electcon        10           10
#> 19:   urbrur    walls    water        10           24
#> 20:   urbrur    walls electcon        10            6
#> 21:   urbrur    water electcon        10           19
#> 22:     roof    walls    water        10           36
#> 23:     roof    walls electcon        10           14
#> 24:     roof    water electcon        10           31
#> 25:    walls    water electcon        10           26
#>         dim1     dim2     dim3 threshold unsafe_cells

## data.frame method / using identification levels
riskyCells(testdata2, keyVars=c(1:6), threshold=20,
  useIdentificationLevel=TRUE, level=c(1,1,2,3,3,5))

#>         dim1     dim2     dim3     dim4 threshold unsafe_cells
#>  1:   urbrur       NA       NA       NA        20            0
#>  2:     roof       NA       NA       NA        20            3
#>  3:    walls       NA       NA       NA        20            1
#>  4:    water       NA       NA       NA        20            7
#>  5: electcon       NA       NA       NA        20            0
#>  6:    relat       NA       NA       NA        20            4
#>  7:   urbrur     roof       NA       NA        20            7
#>  8:   urbrur    walls       NA       NA        20            5
#>  9:   urbrur    water       NA       NA        20           12
#> 10:   urbrur electcon       NA       NA        20            1
#> 11:   urbrur    relat       NA       NA        20            9
#> 12:     roof    walls       NA       NA        20            9
#> 13:     roof    water       NA       NA        20           24
#> 14:     roof electcon       NA       NA        20            8
#> 15:     roof    relat       NA       NA        20           17
#> 16:    walls    water       NA       NA        20           17
#> 17:    walls electcon       NA       NA        20            3
#> 18:    walls    relat       NA       NA        20           11
#> 19:    water electcon       NA       NA        20           13
#> 20:    water    relat       NA       NA        20           26
#> 21: electcon    relat       NA       NA        20            8
#> 22:   urbrur     roof    walls       NA        20           15
#> 23:   urbrur     roof    water       NA        20           32
#> 24:   urbrur     roof electcon       NA        20           13
#> 25:   urbrur     roof    relat       NA        20           26
#> 26:   urbrur    walls    water       NA        20           24
#> 27:   urbrur    walls electcon       NA        20            9
#> 28:   urbrur    walls    relat       NA        20           21
#> 29:   urbrur    water electcon       NA        20           19
#> 30:   urbrur    water    relat       NA        20           36
#> 31:   urbrur electcon    relat       NA        20           15
#> 32:     roof    walls    water       NA        20           36
#> 33:     roof    walls electcon       NA        20           16
#> 34:     roof    walls    relat       NA        20           26
#> 35:     roof    water electcon       NA        20           31
#> 36:     roof    water    relat       NA        20           50
#> 37:     roof electcon    relat       NA        20           24
#> 38:    walls    water electcon       NA        20           26
#> 39:    walls    water    relat       NA        20           41
#> 40:    walls electcon    relat       NA        20           18
#> 41:    water electcon    relat       NA        20           35
#> 42:   urbrur     roof    walls    water        20           47
#> 43:   urbrur     roof    walls electcon        20           23
#> 44:   urbrur     roof    walls    relat        20           37
#> 45:   urbrur     roof    water electcon        20           38
#> 46:   urbrur     roof    water    relat        20           62
#> 47:   urbrur     roof electcon    relat        20           35
#> 48:   urbrur    walls    water electcon        20           34
#> 49:   urbrur    walls    water    relat        20           53
#> 50:   urbrur    walls electcon    relat        20           28
#> 51:   urbrur    water electcon    relat        20           46
#> 52:     roof    walls    water electcon        20           44
#> 53:     roof    walls    water    relat        20           65
#> 54:     roof    walls electcon    relat        20           36
#> 55:     roof    water electcon    relat        20           56
#> 56:    walls    water electcon    relat        20           52
#>         dim1     dim2     dim3     dim4 threshold unsafe_cells

riskyCells(testdata2, keyVars=c(1,3,4,6), threshold=10,
  useIdentificationLevel=TRUE, level=c(1,2,2,4))

#>       dim1  dim2  dim3 threshold unsafe_cells
#>  1: urbrur    NA    NA        10            0
#>  2:  walls    NA    NA        10            1
#>  3:  water    NA    NA        10            4
#>  4:  relat    NA    NA        10            3
#>  5: urbrur walls    NA        10            2
#>  6: urbrur water    NA        10            9
#>  7: urbrur relat    NA        10            7
#>  8:  walls water    NA        10           16
#>  9:  walls relat    NA        10           10
#> 10:  water relat    NA        10           23
#> 11: urbrur walls water        10           24
#> 12: urbrur walls relat        10           18
#> 13: urbrur water relat        10           35
#> 14:  walls water relat        10           41

## sdcMicroObj-method / all combinations up to maxDim
testdata2[1:6] <- lapply(1:6, function(x) {
  testdata2[[x]] <- as.factor(testdata2[[x]])
})
sdc <- createSdcObj(testdata2,
  keyVars=c('urbrur','roof','walls','water','electcon','relat','sex'),
  numVars=c('expend','income','savings'), w='sampling_weight')

r0 <- riskyCells(sdc, useIdentificationLevel=FALSE, threshold=c(20,10,5), maxDim=3)
## in case key-variables have been modified, we get counts for original and modified data
sdc <- groupAndRename(sdc, var="roof", before=c("5","6","9"), after=c("5+"))
r1 <- riskyCells(sdc, useIdentificationLevel=FALSE, threshold=c(10,5,3), maxDim=3)

## sdcMicroObj-method / using identification levels
riskyCells(sdc, useIdentificationLevel=TRUE, threshold=10, level=c(c(1,1,3,4,5,5,5)))

#>         dim1     dim2     dim3     dim4 threshold unsafe_cells_orig
#>  1:   urbrur       NA       NA       NA        10                 0
#>  2:     roof       NA       NA       NA        10                 3
#>  3:    walls       NA       NA       NA        10                 1
#>  4:    water       NA       NA       NA        10                 4
#>  5: electcon       NA       NA       NA        10                 0
#>  6:    relat       NA       NA       NA        10                 3
#>  7:      sex       NA       NA       NA        10                 0
#>  8:   urbrur     roof       NA       NA        10                 6
#>  9:   urbrur    walls       NA       NA        10                 2
#> 10:   urbrur    water       NA       NA        10                 9
#> 11:   urbrur electcon       NA       NA        10                 1
#> 12:   urbrur    relat       NA       NA        10                 7
#> 13:   urbrur      sex       NA       NA        10                 0
#> 14:     roof    walls       NA       NA        10                 6
#> 15:     roof    water       NA       NA        10                24
#> 16:     roof electcon       NA       NA        10                 6
#> 17:     roof    relat       NA       NA        10                16
#> 18:     roof      sex       NA       NA        10                 6
#> 19:    walls    water       NA       NA        10                16
#> 20:    walls electcon       NA       NA        10                 3
#> 21:    walls    relat       NA       NA        10                10
#> 22:    walls      sex       NA       NA        10                 2
#> 23:    water electcon       NA       NA        10                 9
#> 24:    water    relat       NA       NA        10                23
#> 25:    water      sex       NA       NA        10                14
#> 26: electcon    relat       NA       NA        10                 6
#> 27: electcon      sex       NA       NA        10                 0
#> 28:    relat      sex       NA       NA        10                 5
#> 29:   urbrur     roof    walls       NA        10                12
#> 30:   urbrur     roof    water       NA        10                32
#> 31:   urbrur     roof electcon       NA        10                10
#> 32:   urbrur     roof    relat       NA        10                26
#> 33:   urbrur     roof      sex       NA        10                14
#> 34:   urbrur    walls    water       NA        10                24
#> 35:   urbrur    walls electcon       NA        10                 6
#> 36:   urbrur    walls    relat       NA        10                18
#> 37:   urbrur    walls      sex       NA        10                10
#> 38:   urbrur    water electcon       NA        10                19
#> 39:   urbrur    water    relat       NA        10                35
#> 40:   urbrur    water      sex       NA        10                24
#> 41:   urbrur electcon    relat       NA        10                12
#> 42:   urbrur electcon      sex       NA        10                 3
#> 43:   urbrur    relat      sex       NA        10                11
#> 44:     roof    walls    water       NA        10                36
#> 45:     roof    walls electcon       NA        10                14
#> 46:     roof    walls    relat       NA        10                24
#> 47:     roof    walls      sex       NA        10                18
#> 48:     roof    water electcon       NA        10                31
#> 49:     roof    water    relat       NA        10                50
#> 50:     roof    water      sex       NA        10                48
#> 51:     roof electcon    relat       NA        10                24
#> 52:     roof electcon      sex       NA        10                16
#> 53:     roof    relat      sex       NA        10                23
#> 54:    walls    water electcon       NA        10                26
#> 55:    walls    water    relat       NA        10                41
#> 56:    walls    water      sex       NA        10                34
#> 57:    walls electcon    relat       NA        10                14
#> 58:    walls electcon      sex       NA        10                 6
#> 59:    walls    relat      sex       NA        10                14
#> 60:    water electcon    relat       NA        10                34
#> 61:    water electcon      sex       NA        10                25
#> 62:    water    relat      sex       NA        10                36
#> 63: electcon    relat      sex       NA        10                10
#> 64:   urbrur     roof    walls    water        10                47
#> 65:   urbrur     roof    walls electcon        10                22
#> 66:   urbrur     roof    walls    relat        10                37
#> 67:   urbrur     roof    walls      sex        10                30
#> 68:   urbrur     roof    water electcon        10                38
#> 69:   urbrur     roof    water    relat        10                62
#> 70:   urbrur     roof    water      sex        10                63
#> 71:   urbrur     roof electcon    relat        10                34
#> 72:   urbrur     roof electcon      sex        10                25
#> 73:   urbrur     roof    relat      sex        10                36
#> 74:   urbrur    walls    water electcon        10                34
#> 75:   urbrur    walls    water    relat        10                53
#> 76:   urbrur    walls    water      sex        10                48
#> 77:   urbrur    walls electcon    relat        10                27
#> 78:   urbrur    walls electcon      sex        10                19
#> 79:   urbrur    walls    relat      sex        10                28
#> 80:   urbrur    water electcon    relat        10                46
#> 81:   urbrur    water electcon      sex        10                36
#> 82:   urbrur    water    relat      sex        10                48
#> 83:   urbrur electcon    relat      sex        10                24
#> 84:     roof    walls    water electcon        10                44
#> 85:     roof    walls    water    relat        10                65
#> 86:     roof    walls    water      sex        10                72
#> 87:     roof    walls electcon    relat        10                35
#> 88:     roof    walls electcon      sex        10                30
#> 89:     roof    walls    relat      sex        10                36
#> 90:     roof    water electcon    relat        10                56
#> 91:     roof    water electcon      sex        10                57
#> 92:     roof    water    relat      sex        10                65
#> 93:     roof electcon    relat      sex        10                33
#> 94:    walls    water electcon    relat        10                52
#> 95:    walls    water electcon      sex        10                48
#> 96:    walls    water    relat      sex        10                58
#> 97:    walls electcon    relat      sex        10                26
#> 98:    water electcon    relat      sex        10                49
#>         dim1     dim2     dim3     dim4 threshold unsafe_cells_orig
#>     unsafe_cells_manip
#>  1:                  0
#>  2:                  0
#>  3:                  1
#>  4:                  4
#>  5:                  0
#>  6:                  3
#>  7:                  0
#>  8:                  2
#>  9:                  2
#> 10:                  9
#> 11:                  1
#> 12:                  7
#> 13:                  0
#> 14:                  4
#> 15:                 20
#> 16:                  3
#> 17:                 11
#> 18:                  2
#> 19:                 16
#> 20:                  3
#> 21:                 10
#> 22:                  2
#> 23:                  9
#> 24:                 23
#> 25:                 14
#> 26:                  6
#> 27:                  0
#> 28:                  5
#> 29:                 10
#> 30:                 28
#> 31:                  7
#> 32:                 21
#> 33:                  8
#> 34:                 24
#> 35:                  6
#> 36:                 18
#> 37:                 10
#> 38:                 19
#> 39:                 35
#> 40:                 24
#> 41:                 12
#> 42:                  3
#> 43:                 11
#> 44:                 34
#> 45:                 12
#> 46:                 21
#> 47:                 16
#> 48:                 28
#> 49:                 46
#> 50:                 40
#> 51:                 20
#> 52:                 10
#> 53:                 17
#> 54:                 26
#> 55:                 41
#> 56:                 34
#> 57:                 14
#> 58:                  6
#> 59:                 14
#> 60:                 34
#> 61:                 25
#> 62:                 36
#> 63:                 10
#> 64:                 45
#> 65:                 21
#> 66:                 34
#> 67:                 28
#> 68:                 35
#> 69:                 59
#> 70:                 55
#> 71:                 31
#> 72:                 19
#> 73:                 29
#> 74:                 34
#> 75:                 53
#> 76:                 48
#> 77:                 27
#> 78:                 19
#> 79:                 28
#> 80:                 46
#> 81:                 36
#> 82:                 48
#> 83:                 24
#> 84:                 42
#> 85:                 63
#> 86:                 68
#> 87:                 33
#> 88:                 26
#> 89:                 33
#> 90:                 53
#> 91:                 51
#> 92:                 61
#> 93:                 28
#> 94:                 52
#> 95:                 48
#> 96:                 58
#> 97:                 26
#> 98:                 49
#>     unsafe_cells_manip