A real-world data set on household income and expenditures
Source: R/dataSets.R

A concise (1-5 lines) description of the dataset.

Format

testdata: a data frame with 4580 observations on the following 15 variables.

urbrur

a numeric vector

roof

a numeric vector

walls

a numeric vector

water

a numeric vector

electcon

a numeric vector

relat

a numeric vector

sex

a numeric vector

age

a numeric vector

hhcivil

a numeric vector

expend

a numeric vector

income

a numeric vector

savings

a numeric vector

ori_hid

a numeric vector

sampling_weight

a numeric vector

household_weights

a numeric vector

testdata2: A data frame with 93 observations on the following 19 variables.

urbrur

a numeric vector

roof

a numeric vector

walls

a numeric vector

water

a numeric vector

electcon

a numeric vector

relat

a numeric vector

sex

a numeric vector

age

a numeric vector

hhcivil

a numeric vector

expend

a numeric vector

income

a numeric vector

savings

a numeric vector

ori_hid

a numeric vector

sampling_weight

a numeric vector

represent

a numeric vector

category_count

a numeric vector

relat2

a numeric vector

water2

a numeric vector

water3

a numeric vector

References

The International Household Survey Network, www.ihsn.org

Examples

head(testdata)

#>   urbrur roof walls water electcon relat sex age hhcivil   expend   income
#> 1      2    4     3     3        1     1   1  46       2 90929693 57800000
#> 2      2    4     3     3        1     2   2  41       2 27338058 25300000
#> 3      2    4     3     3        1     3   1   9       1 26524717 69200000
#> 4      2    4     3     3        1     3   1   6       1 18073948 79600000
#> 5      2    4     2     3        1     1   1  52       2  6713247 90300000
#> 6      2    4     2     3        1     2   2  47       2 49057636 32900000
#>     savings ori_hid sampling_weight household_weights
#> 1  116258.5       1             100          25.00000
#> 2  279345.0       1             100          25.00000
#> 3 5495381.0       1             100          25.00000
#> 4 8695862.0       1             100          25.00000
#> 5  203620.2       2             100          16.66667
#> 6 1021268.0       2             100          16.66667

head(testdata2)

#>   urbrur roof walls water electcon relat sex age hhcivil   expend   income
#> 1      1    2     3     1        4     1   1  69       1 30772412 72500000
#> 2      1    2     3     3        4     3   1  35       1 77536815 10100000
#> 3      1    2     3     3        4     1   2  80       4 15103477 56700000
#> 4      1    2     3     4        1     3   1  14       1 99118978 12300000
#> 5      1    2     3     4        4     2   2  28       2 55576313 26300000
#> 6      1    4     2     1        1     3   1  25       1 46468920 88600000
#>   savings ori_hid sampling_weight represent category_count relat2 water2 water3
#> 1 1230366     893             100         1              1     10     10    100
#> 2 1856943     621             100         1              1     30     10    100
#> 3 4807111     621             100         1              1     10     10    100
#> 4 7195081     667             100         1             11     30     20    200
#> 5 4934387     658             100         1             20     10     20    200
#> 6 8165375     625             100         1             58     30     10    100